requirements.txt
langchain
numpy
openai
huggingface_hub
streamlit 
langchain 
openai 
faiss-cpu 
pypdf 
python-dotenv 
dateparser

